requirements.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5.3
certifi==2020.6.20
chardet==3.0.4
Django==3.0.6
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
dodgy==0.2.1
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
mypy==0.770
mypy-extensions==0.4.3
pep8-naming==0.4.1
Pillow==7.1.2
prospector==1.2.0
pycodestyle==2.6.0
pydocstyle==5.0.2
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.4.4
pylint-celery==0.3
pylint-django==2.0.12
pylint-flask==0.6
pylint-plugin-utils==0.6
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requirements-detector==0.7
setoptconf==0.2.0
six==1.15.0
snowballstemmer==2.0.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
wrapt==1.11.2
